+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessStateDefinitionManagerTest.scala
@@ -22,6 +22,8 @@ class PeriodicProcessStateDefinitionManagerTest extends AnyFunSuite with Matcher
 
   private val fooRunAt = LocalDateTime.of(2023, 1, 1, 10, 0)
 
+  private val fooCreatedAt = fooRunAt.minusMinutes(5)
+
   private val notNamedScheduleId = ScheduleId(fooProcessId, ScheduleName(None))
 
   private val nextDeploymentId = new AtomicLong()
@@ -32,6 +34,7 @@ class PeriodicProcessStateDefinitionManagerTest extends AnyFunSuite with Matcher
     val deploymentStatus = DeploymentStatus(
       generateDeploymentId,
       notNamedScheduleId,
+      fooCreatedAt,
       fooRunAt,
       PeriodicProcessDeploymentStatus.Scheduled,
       processActive = true,
@@ -41,13 +44,14 @@ class PeriodicProcessStateDefinitionManagerTest extends AnyFunSuite with Matcher
     statusTooltip(status) shouldEqual "Scheduled at: 2023-01-01 10:00 status: Scheduled"
   }
 
-  test("display periodic deployment status for named schedules") {
+  test("display sorted periodic deployment status for named schedules") {
     val firstScheduleId = generateScheduleId
     val firstDeploymentStatus = DeploymentStatus(
       generateDeploymentId,
       firstScheduleId,
+      fooCreatedAt.minusMinutes(1),
       fooRunAt,
-      PeriodicProcessDeploymentStatus.Scheduled,
+      PeriodicProcessDeploymentStatus.Deployed,
       processActive = true,
       None
     )
@@ -55,15 +59,16 @@ class PeriodicProcessStateDefinitionManagerTest extends AnyFunSuite with Matcher
     val secDeploymentStatus = DeploymentStatus(
       generateDeploymentId,
       secScheduleId,
+      fooCreatedAt,
       fooRunAt,
-      PeriodicProcessDeploymentStatus.Deployed,
+      PeriodicProcessDeploymentStatus.Scheduled,
       processActive = true,
       None
     )
     val status = PeriodicProcessStatus(List(firstDeploymentStatus, secDeploymentStatus), List.empty)
     statusTooltip(status) shouldEqual
-      s"""Schedule ${firstScheduleId.scheduleName.display} scheduled at: 2023-01-01 10:00 status: Scheduled,
-         |Schedule ${secScheduleId.scheduleName.display} scheduled at: 2023-01-01 10:00 status: Deployed""".stripMargin
+      s"""Schedule ${secScheduleId.scheduleName.display} scheduled at: 2023-01-01 10:00 status: Scheduled,
+         |Schedule ${firstScheduleId.scheduleName.display} scheduled at: 2023-01-01 10:00 status: Deployed""".stripMargin
   }
 
   private def generateDeploymentId = PeriodicProcessDeploymentId(nextDeploymentId.getAndIncrement())